requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
dill==0.3.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
multiprocess==0.70.10
numpy==1.19.2
opencv-python==4.4.0.44
pathos==0.2.6
Pillow==8.0.0
pox==0.2.8
ppft==1.6.6.2
psutil==5.7.2
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
six==1.15.0
tornado==6.0.4
tzlocal==2.1
